edit api: convert to tsx
